@@ -12,7 +12,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/asticode/go-bindata"
+	"github.com/firebat20/go-bindata"
 )
 
 func main() {
@@ -88,9 +88,9 @@ func parseArgs() *bindata.Config {
 // parseRecursive determines whether the given path has a recrusive indicator and
 // returns a new path with the recursive indicator chopped off if it does.
 //
-//  ex:
-//      /path/to/foo/...    -> (/path/to/foo, true)
-//      /path/to/bar        -> (/path/to/bar, false)
+//	ex:
+//	    /path/to/foo/...    -> (/path/to/foo, true)
+//	    /path/to/bar        -> (/path/to/bar, false)
 func parseInput(path string) bindata.InputConfig {
 	if strings.HasSuffix(path, "/...") {
 		return bindata.InputConfig{
